+++ python/cusignal/filtering/resample.py
@@ -84,13 +84,7 @@ def _design_resample_poly(up, down, window):
 
 
 def decimate(
-    x,
-    q,
-    n=None,
-    axis=-1,
-    zero_phase=True,
-    cp_stream=cp.cuda.stream.Stream.null,
-    autosync=True,
+    x, q, n=None, axis=-1, zero_phase=True,
 ):
     """
     Downsample the signal after applying an anti-aliasing filter.
@@ -110,17 +104,6 @@ def decimate(
         Prevent shifting the outputs back by the filter's
         group delay when using an FIR filter. The default value of ``True`` is
         recommended, since a phase shift is generally not desired.
-    cp_stream : CuPy stream, optional
-        Option allows upfirdn to run in a non-default stream. The use
-        of multiple non-default streams allow multiple kernels to
-        run concurrently. Default is cp.cuda.stream.Stream.null
-        or default stream.
-    autosync : bool, optional
-        Option to automatically synchronize cp_stream. This will block
-        the host code until kernel is finished on the GPU. Setting to
-        false will allow asynchronous operation but might required
-        manual synchronize later `cp_stream.synchronize()`.
-        Default is True.
 
     Returns
     -------
@@ -153,7 +136,7 @@ def decimate(
         # upfirdn is generally faster than lfilter by a factor equal to the
         # downsampling factor, since it only calculates the needed outputs
         n_out = x.shape[axis] // q + bool(x.shape[axis] % q)
-        y = upfirdn(b, x, 1, q, axis, cp_stream, autosync)
+        y = upfirdn(b, x, 1, q, axis)
         sl[axis] = slice(None, n_out, None)
 
     return y[tuple(sl)]
@@ -293,13 +276,7 @@ def resample(x, num, t=None, axis=0, window=None, domain="time"):
 
 
 def resample_poly(
-    x,
-    up,
-    down,
-    axis=0,
-    window=("kaiser", 5.0),
-    cp_stream=cp.cuda.stream.Stream.null,
-    autosync=True,
+    x, up, down, axis=0, window=("kaiser", 5.0),
 ):
     """
     Resample `x` along the given axis using polyphase filtering.
@@ -323,17 +300,6 @@ def resample_poly(
     window : string, tuple, or array_like, optional
         Desired window to use to design the low-pass filter, or the FIR filter
         coefficients to employ. See below for details.
-    cp_stream : CuPy stream, optional
-        Option allows upfirdn to run in a non-default stream. The use
-        of multiple non-default streams allow multiple kernels to
-        run concurrently. Default is cp.cuda.stream.Stream.null
-        or default stream.
-    autosync : bool, optional
-        Option to automatically synchronize cp_stream. This will block
-        the host code until kernel is finished on the GPU. Setting to
-        false will allow asynchronous operation but might required
-        manual synchronize later `cp_stream.synchronize()`.
-        Default is True.
 
     Returns
     -------
@@ -394,6 +360,7 @@ def resample_poly(
     >>> plt.legend(['resample', 'resamp_poly', 'data'], loc='best')
     >>> plt.show()
     """
+
     x = asarray(x)
     up = int(up)
     down = int(down)
@@ -438,20 +405,15 @@ def resample_poly(
     n_pre_remove_end = n_pre_remove + n_out
 
     # filter then remove excess
-    y = upfirdn(h, x, up, down, axis, cp_stream, autosync)
+    y = upfirdn(h, x, up, down, axis)
     keep = [slice(None)] * x.ndim
     keep[axis] = slice(n_pre_remove, n_pre_remove_end)
+
     return y[tuple(keep)]
 
 
 def upfirdn(
-    h,
-    x,
-    up=1,
-    down=1,
-    axis=-1,
-    cp_stream=cp.cuda.stream.Stream.null,
-    autosync=True,
+    h, x, up=1, down=1, axis=-1,
 ):
     """
     Upsample, FIR filter, and downsample
@@ -470,17 +432,6 @@ def upfirdn(
         The axis of the input data array along which to apply the
         linear filter. The filter is applied to each subarray along
         this axis. Default is -1.
-    cp_stream : CuPy stream, optional
-        Option allows upfirdn to run in a non-default stream. The use
-        of multiple non-default streams allow multiple kernels to
-        run concurrently. Default is cp.cuda.stream.Stream.null
-        or default stream.
-    autosync : bool, optional
-        Option to automatically synchronize cp_stream. This will block
-        the host code until kernel is finished on the GPU. Setting to
-        false will allow asynchronous operation but might required
-        manual synchronize later `cp_stream.synchronize()`.
-        Default is True.
 
     Returns
     -------
@@ -541,9 +492,8 @@ def upfirdn(
            [ 6.,  7.],
            [ 6.,  7.]])
     """
+
     x = cp.asarray(x)
     ufd = _UpFIRDn(h, x.dtype, up, down)
     # This is equivalent to (but faster than) using cp.apply_along_axis
-    return ufd.apply_filter(
-        x, axis, cp_stream=cp_stream, autosync=autosync
-    )
+    return ufd.apply_filter(x, axis)